Fix performance issue when saving snapshot to disk
This was causing congestion in `file_server_2` mailbox, because each
snapshot save would use `read_file_info` function when making sure that
the directory exists before saving the file. Changing the code so the
directory is only created if writing the snapshot fails (when the
directory doesn't exist yet) improves the performance significantly.